Get rid of libpq-fe.h inclusion in public headers
The problem with libpq-fe.h is that it is installed in unpredictable
places on different platforms. As a result, a user that uses ODB with
PostgreSQL (and who doesn't really know or care about libpq) is forced
to make sure their application is able to find and include libpq-fe.h
correctly.

Luckily for us, we only use a handful of libpq pointers in public
headers and the workaround is to forward declare them and use that
instead of including libpq-fe.h (which is instead included in source
files).