Correct the LOB callback aspects of the bind mechanism
The parameter callback context must be const qualified. This is both sensible
and necessary to avoid unecessary casts in the generated code during image
manipulation. The result callback context must not be const qualified as the
result callback should be permitted to modify it.

Both the callback and context members of the bind structure must be pointers
to the relevant union types in order for the binding mechanism to work
correctly.